+++ doc/release-notes.rst
@@ -63,6 +63,10 @@ This release is scheduled for May/June 2024. You can already check out a `previe
 
 - | 16, 17: New mapping for odoo reorder points, which uses the new buffer.maximum field.
 
+- | The "odoo export" command now sends back information on the planned delivery date of every
+    open sales order.
+  | This can be valuable feedback in odoo to the sales team.
+
 .. rubric:: System administration
 
 - | Ability to customize the "export plan results" task.